requirements.txt
torch-ac>=1.4.0
minigrid
tensorboardX>=1.6
numpy>=1.3
gymnasium>=0.26
faiss-cpu>=1.6.3
torch
wandb


